+++ ChangeLog.md
@@ -1,3 +1,5 @@
 # Changelog for WaveCut
 
+- Add new Intends: TeacherDirections and Pronunciation
+
 ## Unreleased changes